No secondary environments.
The only expression where a secondary environment was created, was
LetExpr. This was fixed, because it allowed unintentional self-recursive
variables (e.g. fibonacci-function). Now a let statement opens a
secondary environment to assign the variables, then lambda expressions,
which have a substition right after them, are created to the respective
variables in the secondary environment, this allows us to close the
environment again and handle the let expression with the main
environment.

Example:

    > let x = 0 in x
    (\x = x) 0
    0